+++ frontend/src/App.js
@@ -12,7 +12,7 @@ import GroupsList from "./pages/Groups/GroupsList/GroupsList";
 import ReportArtwork from "./pages/ReportArtwork/ReportArtwork";
 import TabBar from "./components/TabBar/TabBar";
 import useProfileInfo from "./hooks/useProfileInfo";
-import NewAccount from "./pages/NewAccount/NewAccount";
+import NewAccount from "./pages/SignUp/SignUp";
 import { useEffect } from "react";
 
 const userId = "VXNlclR5cGU6am9obkBqb2huLmNvbQ==";
@@ -46,7 +46,7 @@ function App() {
             <Route path="/groups" component={GroupsList} />
             <Route path="/group/:id" component={GroupPage} />
             <Route path="/style-guide" component={StyleGuide} />
-            <Route path="/new-account" component={NewAccount} />
+            <Route path="/sign-up" component={NewAccount} />
             <Route path="*" component={ArtMap} />
           </Switch>
         </div>
