+++ frontend/src/pages/SignUp/SignUp.scss
